+++ rembish_org/__init__.py
@@ -1,2 +1,0 @@
-from .application import create_app
-from .version import __version__