+++ rembish_org/libraries/authz.py
@@ -1,8 +1,0 @@
-from flask_security import SQLAlchemyUserDatastore, Security
-
-from .database import db
-from ..models.role import Role
-from ..models.user import User
-
-datastore = SQLAlchemyUserDatastore(db, User, Role)
-security = Security()